Tools/flashtool/requirements.txt
# pip install -r requirements.txt
#

# COBS needs a working MSVC installation on windows. 
# Until that's fixed, I've included a copy of cobs.py.
#
# cobs=1.0.0
#

pyserial==2.7
crcmod==1.7

--allow-external intelhex
--allow-unverified intelhex
intelhex==1.5

PyOpenGL

# Requirements for matplotlib
#
python-dateutil
pyparsing

